Updated TODO in main.js
On a side note, I have reached critical mass in the amount of content I
want to put in the game, that means everything has been more or less
design. Past this point it makes no sense to add more stuff than what is
already written down. That is good.